% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{browse_colorpile}
\alias{browse_colorpile}
\title{Browse colorpile palettes}
\usage{
browse_colorpile(pal)
}
\arguments{
\item{pal}{Character vector of palette names}
}
\description{
Browse colorpile palettes
}

