% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette.R
\name{create_palette}
\alias{create_palette}
\title{Create palette for farbenfroh}
\usage{
create_palette(name, authors, colors = NULL, type = NULL,
  github_user = NULL, description = Sys.Date(), keywords, date)
}
\arguments{
\item{name}{The short name of the colour scheme}

\item{authors}{vector of name(s) of palette author(s)}

\item{colors}{Vector of colors, each in hex format}

\item{type}{One of "divergent", "qualitative", "sequential", "other"}

\item{github_user}{The GitHub username of the palette's author (or primary author if there is more than one name in authors)}

\item{description}{A description, possibly longer than the name}

\item{keywords}{Vector of palette keywords}

\item{date}{Date scheme was created or modified}
}
\description{
create palette for farbenfroh
}

