% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate.R
\name{palette_pply}
\alias{palette_pply}
\title{Apply function over palettes}
\usage{
palette_pply(repo_dir, fun, ...)
}
\arguments{
\item{repo_dir}{directory of palette repository}

\item{fun}{function to apply over files}
}
\description{
Apply function over palettes
}

