% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_colorpile}
\alias{scale_color_colorpile}
\alias{scale_colour_colorpile}
\alias{scale_fill_colorpile}
\title{colorpile colour palettes for ggplot2}
\usage{
scale_colour_colorpile(name, ...)

scale_color_colorpile(name, ...)

scale_fill_colorpile(name, ...)
}
\arguments{
\item{name}{Name of palette.}

\item{...}{Arguments passed on to discrete_scale to control name, limits,
breaks, labels and so forth.}
}
\description{
colorpile colour palettes for ggplot2
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)
  ggplot(mtcars, aes(x = mpg, y = hp, colour = factor(cyl))) +
    geom_point() +
    scale_colour_colorpile("Royal1")
}
}

