% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farben.R
\name{scale_colour_farben}
\alias{scale_color_farben}
\alias{scale_colour_farben}
\alias{scale_fill_farben}
\title{farben colour palettes for ggplot2}
\usage{
scale_colour_farben(name, ...)

scale_color_farben(name, ...)

scale_fill_farben(name, ...)
}
\arguments{
\item{name}{Name of palette.}

\item{...}{Arguments passed on to discrete_scale to control name, limits,
breaks, labels and so forth.}
}
\description{
farben colour palettes for ggplot2
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)
  ggplot(mtcars, aes(x = mpg, y = hp, colour = factor(cyl))) +
    geom_point() +
    scale_colour_farben("Royal1")
}
}

