% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate.R
\name{search_colorpile}
\alias{search_colorpile}
\title{Search colorpile palettes}
\usage{
search_colorpile(authors = NULL, github_user = NULL, type = NULL)
}
\arguments{
\item{authors}{Character vector of author(s)}

\item{github_user}{String of author's GitHub username}

\item{type}{One of 'divergent', 'qualitative', 'sequential', 'other'}
}
\value{
Character vector of palette names
}
\description{
Filters the palettes in colorpile down to those that match search terms
}
\examples{
search_colorpile(authors = c("Karthik Ram", "Ethan Schoonover"))
search_colorpile(authors = "Cynthia Brewer", type = "diverging")
}

