% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farben.R
\name{create_palette}
\alias{create_palette}
\title{create palette for farbenfroh}
\usage{
create_palette(name, author, github_user = "", description = "",
  keywords = "", colors, type)
}
\arguments{
\item{name}{String for palette name (non-zero length alpha-numeric)}

\item{author}{Character vector of name(s) of palette author(s)}

\item{github_user}{String for palette's author's GitHub username}

\item{description}{String for palette's description}

\item{keywords}{Character vector of palette keywords}

\item{colors}{Character vector of palette colors}

\item{type}{One of "divergent", "qualitative", "sequential", "other"}
}
\description{
create palette for farbenfroh
}

