% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farben.R
\name{farben_palette}
\alias{farben_palette}
\title{solarized colour palette}
\usage{
farben_palette(palette_name)
}
\arguments{
\item{num_values}{Number of values to colour. If more than 9, colours start
to repeat.}
}
\description{
Provides a colour scheme based on the
\href{http://ethanschoonover.com/solarized}{solarized} accent colours.
}
\examples{
if (requireNamespace("scales", quietly = TRUE)) {
  library(scales)
  show_col(farben_palette("example"))
}
}

