% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farben.R
\name{farben_palette}
\alias{farben_palette}
\title{farben_palette}
\usage{
farben_palette(palette_name)
}
\arguments{
\item{palette_name}{Name of palette.}
}
\description{
farben_palette
}
\examples{
if (requireNamespace("scales", quietly = TRUE)) {
  library(scales)
  show_col(farben_palette("wes_a_fave")(9))
}
}

